Multiple changes ...
Add a rudimentary network stack
-------------------------------

based on nats and Jetstream, allows to retrieve current game and
statuses after-the-fact if need be.

Revamp Game around errgroup and channels
----------------------------------------
two chans of enums are use to pass events
from and to network, and also to manage the player's actions to the GUI
With network-confirmed validation of words, the `validationLevel` struct
is used to track the state of each word of ours.